fixes warning and nullability issues

+++ source/ios/AdaptiveCards/AdaptiveCards/AdaptiveCards/ACRView.mm
@@ -523,7 +523,7 @@ typedef UIImage * (^ImageLoadBlock)(NSURL *url);
             ACOBaseCardElement *baseCardElement = _imageContextMap[key];
             if (baseCardElement) {
                 ACRRegistration *reg = [ACRRegistration getInstance];
-                ACRBaseCardElementRenderer<ACRIKVONotificationHandler> *renderer = (ACRBaseCardElementRenderer<ACRIKVONotificationHandler> *)[reg getRenderer:[NSNumber numberWithInt:baseCardElement.type]];
+                ACRBaseCardElementRenderer<ACRIKVONotificationHandler> *renderer = (ACRBaseCardElementRenderer<ACRIKVONotificationHandler> *)[reg getRenderer:[NSNumber numberWithInt:static_cast<int>(baseCardElement.type)]];
                 if (renderer && [[renderer class] conformsToProtocol:@protocol(ACRIKVONotificationHandler)]) {
                     // remove observer early in case background image must be changed to handle mode = repeat
                     [self removeObserver:self forKeyPath:path onObject:object];
